SwiftShip Express/requirements.txt
requests
python-dotenv
pandas
pyarrow
supabase
